+++ src/apps/ScopeChangeRequest/WorkspaceConfig/Garden/gardenConfig.ts
@@ -1,6 +1,6 @@
-import { GardenOptions } from '../../../components/ParkView/Models/gardenOptions';
-import { ScopeChangeItemView } from '../Garden/ScopeChangeGardenItem';
-import { ScopeChangeRequest } from '../Types/scopeChangeRequest';
+import { GardenOptions } from '../../../../components/ParkView/Models/gardenOptions';
+import { ScopeChangeItemView } from './ScopeChangeGardenItem';
+import { ScopeChangeRequest } from '../../Types/scopeChangeRequest';
 
 export const gardenConfig: GardenOptions<ScopeChangeRequest> = {
     gardenKey: 'originSource',
